rename #40: 잘못된 폴더명 수정
