+++ fe/src/components/common/ProductList/index.tsx
@@ -2,8 +2,8 @@ import { Fragment, useEffect, useRef, useState } from 'react';
 
 import { END_POINT } from '@Constants/endpoint';
 
-import usePullToRefresh from '@Hooks/useFullToRefresh';
 import useInfiniteScroll from '@Hooks/useInfiniteScroll';
+import usePullToRefresh from '@Hooks/usePullToRefresh';
 
 import { Product, ProductResponseData } from '@Types/index';
